高度な設定→追加CSSクラスが追加されない問題を修正
