+++ src/pr-content/component-fontawesome.js
@@ -1,4 +1,4 @@
-import React from 'react';
+        import React from 'react';
 
 export class Fontawesome extends React.Component {
