+++ src/pr-content/component.js
@@ -5,6 +5,7 @@ const {__} = wp.i18n; // Import __() from wp.i18n
 const {Button} = wp.components;
 const {MediaUpload} = wp.editor;
 let {RichText} = wp.editor;
+const {Fragment} = wp.element;
 
 export class Component extends React.Component {
 
@@ -24,23 +25,13 @@ export class Component extends React.Component {
             buttonTarget,
             Image,
             ImageBorderColor,
-            layout,
-            fontAwesomeIconBefore,
-            fontAwesomeIconAfter
         } = attributes;
         let setAttributes = this.props.setAttributes;
         let for_ = this.props.for_;
-        let containerClass = 'vk_prContent';
         let btnClass = 'vk_button';
         let aClass = 'btn btn-block vk_button_link vk_prContent_colTxt_btn';
         let aStyle = {};
         let imageBorderProperty = '';
-
-        if (layout === 'right') {
-            containerClass = `${containerClass} vk_prContent-layout-imageRight`;
-        } else {
-            containerClass = `${containerClass} vk_prContent-layout-imageLeft`;
-        }
 
         if (buttonColorCustom) {
             btnClass = `${btnClass} vk_button-color-custom`;
@@ -87,88 +78,88 @@ export class Component extends React.Component {
 
 
         return (
-            <div className={containerClass}>
-                        <div className="col-sm-6 vk_prContent_colImg">
-                            {for_ === 'edit' ?
-                                <MediaUpload
-                                    onSelect={(value) => setAttributes({Image: value.sizes.full.url})}
-                                    type=" image"
-                                    value={Image}
-                                    render={({open}) => (
-                                        <Button
-                                            onClick={open}
-                                            className={Image ? 'image-button' : 'button button-large'}
-                                        >
-                                            {!Image ? __('Select image', 'vk-blocks') :
-                                                <img
-                                                    className={'vk_prContent_colImg_image'}
-                                                    src={Image}
-                                                    alt={__('Upload image', 'vk-blocks')}
-                                                    style={{border: imageBorderProperty}}
-                                                />}
-                                        </Button>
-                                    )}
+            <Fragment>
+                    <div className="col-sm-6 vk_prContent_colImg">
+                        {for_ === 'edit' ?
+                            <MediaUpload
+                                onSelect={(value) => setAttributes({Image: value.sizes.full.url})}
+                                type=" image"
+                                value={Image}
+                                render={({open}) => (
+                                    <Button
+                                        onClick={open}
+                                        className={Image ? 'image-button' : 'button button-large'}
+                                    >
+                                        {!Image ? __('Select image', 'vk-blocks') :
+                                            <img
+                                                className={'vk_prContent_colImg_image'}
+                                                src={Image}
+                                                alt={__('Upload image', 'vk-blocks')}
+                                                style={{border: imageBorderProperty}}
+                                            />}
+                                    </Button>
+                                )}
+                            />
+                            :
+                            !Image ? __('Select image', 'vk-blocks') :
+                                <img
+                                    className={'vk_prContent_colImg_image'}
+                                    src={Image}
+                                    alt={__('Upload image', 'vk-blocks')}
+                                    style={{border: imageBorderProperty}}
                                 />
-                                :
-                                !Image ? __('Select image', 'vk-blocks') :
-                                    <img
-                                        className={'vk_prContent_colImg_image'}
-                                        src={Image}
-                                        alt={__('Upload image', 'vk-blocks')}
-                                        style={{border: imageBorderProperty}}
-                                    />
-                            }
-                        </div>
-                        <div className="col-sm-6 vk_prContent_colTxt">
-                            {
-                                (() => {
-                                    if (for_ === 'edit') {
-                                        return (
-                                            <React.Fragment>
-                                                <RichText
-                                                    tagName="h3"
-                                                    className={'vk_prContent_colTxt_title'}
-                                                    onChange={(value) => setAttributes({title: value})}
-                                                    value={title}
-                                                    placeholder={__('Input title.', 'vk-blocks')}
-                                                    style={{color: titleColor}}
-                                                />
-                                                < RichText
-                                                    tagName="p"
-																										className={'vk_prContent_colTxt_text'}
-                                                    onChange={(value) => setAttributes({content: value})}
-                                                    value={content}
-                                                    placeholder={__('Input content.', 'vk-blocks')}
-                                                    style={{color: contentColor}}
-                                                />
-                                            </React.Fragment>
-                                        );
-                                    } else {
-                                        return (
-                                            <React.Fragment>
-                                                <RichText.Content
-                                                    tagName="h3"
-                                                    value={title}
-                                                    className={'vk_prContent_colTxt_title'}
-                                                    style={{color: titleColor}}
-                                                />
-                                                <RichText.Content
-                                                    tagName="p"
-																										className={'vk_prContent_colTxt_text'}
-                                                    value={content}
-                                                    style={{color: contentColor}}
-                                                />
-                                            </React.Fragment>);
-                                    }
-                                })()
-                            }
-                            {
+                        }
+                    </div>
+                    <div className="col-sm-6 vk_prContent_colTxt">
+                        {
+                            (() => {
+                                if (for_ === 'edit') {
+                                    return (
+                                        <React.Fragment>
+                                            <RichText
+                                                tagName="h3"
+                                                className={'vk_prContent_colTxt_title'}
+                                                onChange={(value) => setAttributes({title: value})}
+                                                value={title}
+                                                placeholder={__('Input title.', 'vk-blocks')}
+                                                style={{color: titleColor}}
+                                            />
+                                            < RichText
+                                                tagName="p"
+                                                className={'vk_prContent_colTxt_text'}
+                                                onChange={(value) => setAttributes({content: value})}
+                                                value={content}
+                                                placeholder={__('Input content.', 'vk-blocks')}
+                                                style={{color: contentColor}}
+                                            />
+                                        </React.Fragment>
+                                    );
+                                } else {
+                                    return (
+                                        <React.Fragment>
+                                            <RichText.Content
+                                                tagName="h3"
+                                                value={title}
+                                                className={'vk_prContent_colTxt_title'}
+                                                style={{color: titleColor}}
+                                            />
+                                            <RichText.Content
+                                                tagName="p"
+                                                className={'vk_prContent_colTxt_text'}
+                                                value={content}
+                                                style={{color: contentColor}}
+                                            />
+                                        </React.Fragment>);
+                                }
+                            })()
+                        }
+                        {
 
-                                //ボタンテキストが入力されるとボタンを表示。
-                                (() => {
-                                    if (buttonText !== '' && buttonText !== undefined ) {
-                                        return (
-																					<div className={btnClass}>
+                            //ボタンテキストが入力されるとボタンを表示。
+                            (() => {
+                                if (buttonText !== '' && buttonText !== undefined ) {
+                                    return (
+                                        <div className={btnClass}>
                                             <a href={url}
                                                className={aClass}
                                                target={buttonTarget? '_blank':null}
@@ -178,13 +169,13 @@ export class Component extends React.Component {
                                                     attributes={attributes}
                                                 />
                                             </a>
-																					</div>
-                                        );
-                                    }
-                                })()
-                            }
-                        </div>
-            </div>
+                                        </div>
+                                    );
+                                }
+                            })()
+                        }
+                    </div>
+            </Fragment>
         );
     }
 }
