+++ src/pr-content/deprecated/component-fontawesome.js
